Dockerfile Update
This commit includes several changes which would make TEToolkit able to produce a docker container for dockerhub with a manual github action, however to do so would require hammell lab to create a docker account, and add Docker secrets to the TEtranscript repo. It would also the require that the .github/workflows action be updated with versioning information, and the user id of the docker hub account.